Feat: add account verification by email

+++ demo/src/main/java/com/spring_security/jwt_auth/demo/security/authentication/AuthenticationService.java
@@ -1,5 +1,6 @@
 package com.spring_security.jwt_auth.demo.security.authentication;
 
+import com.spring_security.jwt_auth.demo.exception.AccountNotVerifiedException;
 import com.spring_security.jwt_auth.demo.exception.InvalidRefreshTokenException;
 import com.spring_security.jwt_auth.demo.model.User;
 import com.spring_security.jwt_auth.demo.security.dto.request.AuthReq;
@@ -9,10 +10,12 @@ import com.spring_security.jwt_auth.demo.security.dto.response.AuthRes;
 import com.spring_security.jwt_auth.demo.security.jwt.JwtService;
 import com.spring_security.jwt_auth.demo.security.token.RefreshToken;
 import com.spring_security.jwt_auth.demo.security.token.RefreshTokenRepository;
+import com.spring_security.jwt_auth.demo.service.EmailService;
 import com.spring_security.jwt_auth.demo.service.UserService;
 import jakarta.servlet.http.HttpServletRequest;
 import java.security.Principal;
 import lombok.RequiredArgsConstructor;
+import org.springframework.scheduling.annotation.Async;
 import org.springframework.security.authentication.AuthenticationProvider;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
@@ -26,6 +29,7 @@ public class AuthenticationService {
   private final AuthenticationProvider authenticationProvider;
   private final UserService userService;
   private final RefreshTokenRepository tokenRepository;
+  private final EmailService emailService;
 
   @Transactional
   public AuthRes authenticate(AuthReq authReq) {
@@ -33,14 +37,17 @@ public class AuthenticationService {
     UsernamePasswordAuthenticationToken authentication = new UsernamePasswordAuthenticationToken(authReq.getEmail(), authReq.getPassword());
     Authentication auth = authenticationProvider.authenticate(authentication);
     User user = (User) auth.getPrincipal();
+    if(!user.isEnabled())
+      throw new AccountNotVerifiedException("Account not verified");
 
     tokenRepository.deleteByUser(user);
     return getAuthRes(user);
   }
 
-  public AuthRes register(RegisterReq registerReq) {
+  public String register(RegisterReq registerReq){
     User user = userService.createUser(registerReq);
-    return getAuthRes(user);
+    emailService.sendConfirmationEmail(user);
+    return "Verification code sent";
   }
 
 
@@ -83,5 +90,10 @@ public class AuthenticationService {
         .build());
 
     return new AuthRes(accessToken, refreshToken);
+  }
+
+  public String verifyAccount(String token) {
+    userService.enableUser(jwtService.extractUsername(token));
+    return "Account verified";
   }
 }
